Moved all applications to corresponding files
