+++ main.py
@@ -5,7 +5,7 @@ from flask import request, Flask, redirect, url_for
 
 from app import (alert_handle, queue_handle, recreate_queue, Incidents, create_or_load_incidents, generate_webhooks,
                  generate_route, buttons_handler)
-from app.im import get_application
+from app.im.helpers import get_application
 from config import settings, check_updates
 
 app = Flask(__name__)
